CLN: Refactor string methods and add PandasObject
Previous PandasObject becomes PandasContainer. New PandasObject becomes
baseclass for more elements (like Index, Categorical, etc.), moves
string methods to baseclass and subclassing objects need only define
`__unicode__` methods to get all string methods for free (and Py2/3
compatible).

CLN: Cleanup extraneous str methods from Panel

CLN: Remove unnecessary string methods from frame

CLN: Change name of TestPandasObjects --> TestPandasContainer
